feature: add image store page
