+++ src/components/index.ts
@@ -1,3 +1,5 @@
 export * from "./FlowPanel";
 export * from "./ImageStore";
+export * from "./ImageShowcase";
+export * from "./ScreenshotButton";
 export * from "./dialog";
